backend/requirements.txt
Django==5.0.10
djangorestframework==3.15.2
psycopg[binary]==3.2.1
django-environ==0.11.2
channels==4.1.0
channels-redis==4.2.0

PyJWT==2.9.0

daphne==4.1.2
